backend/identity-service/main.py
# -*- coding: utf-8 -*-
"""
SVYD LEDGER - Identity Service
Servicio de Identidad: Gestion de usuarios, roles, permisos y autenticacion OAuth 2.0/OIDC
Python + FastAPI
"""

from fastapi import FastAPI, Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, EmailStr, Field
from typing import Optional, List
from datetime import datetime, timedelta
import jwt
import os
from passlib.context import CryptContext

# Configuracion
API_VERSION = "v1"
SERVICE_NAME = "Identity Service"
REGION = "centralus"  # Region parametrizada Central US

app = FastAPI(
    title=f"SVYD LEDGER - {SERVICE_NAME}",
    description="Servicio de Identidad y Autenticacion",
    version=API_VERSION,
    docs_url=f"/api/{API_VERSION}/docs",
    redoc_url=f"/api/{API_VERSION}/redoc"
)

# CORS configurado para produccion
app.add_middleware(
    CORSMiddleware,
    allow_origins=["https://www.svydledger.com", "https://svydledger.com"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Seguridad
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
oauth2_scheme = OAuth2PasswordBearer(tokenUrl=f"/api/{API_VERSION}/auth/token")

# Configuracion JWT (en produccion usar Azure Key Vault)
SECRET_KEY = os.getenv("JWT_SECRET_KEY", "CHANGE_THIS_IN_PRODUCTION")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

# Modelos Pydantic
class UserBase(BaseModel):
    email: EmailStr
    username: str = Field(..., min_length=3, max_length=50)
    full_name: str = Field(..., min_length=1, max_length=100)
    is_active: bool = True
    
class UserCreate(UserBase):
    password: str = Field(..., min_length=8)
    
class User(UserBase):
    id: int
    created_at: datetime
    updated_at: datetime
    roles: List[str] = []
    
    class Config:
        from_attributes = True

class Token(BaseModel):
    access_token: str
    token_type: str
    expires_in: int

class TokenData(BaseModel):
    user_id: Optional[int] = None
    username: Optional[str] = None
    roles: List[str] = []

class Role(BaseModel):
    id: int
    name: str
    description: str
    permissions: List[str] = []

# Funciones de utilidad
def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verificar contrasena"""
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password: str) -> str:
    """Hash de contrasena"""
    return pwd_context.hash(password)

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    """Crear token JWT"""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

async def get_current_user(token: str = Depends(oauth2_scheme)) -> TokenData:
    """Obtener usuario actual del token"""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        user_id: int = payload.get("sub")
        username: str = payload.get("username")
        roles: List[str] = payload.get("roles", [])
        if user_id is None:
            raise credentials_exception
        token_data = TokenData(user_id=user_id, username=username, roles=roles)
    except jwt.PyJWTError:
        raise credentials_exception
    return token_data

# Health Check
@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {
        "status": "healthy",
        "service": SERVICE_NAME,
        "version": API_VERSION,
        "region": REGION,
        "timestamp": datetime.utcnow().isoformat()
    }

# Endpoints de Autenticacion
@app.post(f"/api/{API_VERSION}/auth/register", response_model=User, status_code=status.HTTP_201_CREATED)
async def register(user: UserCreate):
    """
    Registrar nuevo usuario
    TODO: Integrar con Azure SQL Hyperscale
    """
    # TODO: Verificar si el usuario ya existe en la base de datos
    # TODO: Guardar usuario en Azure SQL Hyperscale
    hashed_password = get_password_hash(user.password)
    
    return User(
        id=1,
        email=user.email,
        username=user.username,
        full_name=user.full_name,
        is_active=user.is_active,
        created_at=datetime.utcnow(),
        updated_at=datetime.utcnow(),
        roles=["user"]
    )

@app.post(f"/api/{API_VERSION}/auth/token", response_model=Token)
async def login(form_data: OAuth2PasswordRequestForm = Depends()):
    """
    Login y obtener token JWT
    TODO: Integrar con Azure SQL Hyperscale
    """
    # TODO: Buscar usuario en Azure SQL Hyperscale
    # TODO: Verificar contrasena
    # Por ahora, simulamos la autenticacion
    
    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    access_token = create_access_token(
        data={"sub": 1, "username": form_data.username, "roles": ["user"]},
        expires_delta=access_token_expires
    )
    
    return Token(
        access_token=access_token,
        token_type="bearer",
        expires_in=ACCESS_TOKEN_EXPIRE_MINUTES * 60
    )

@app.get(f"/api/{API_VERSION}/auth/me", response_model=User)
async def get_current_user_info(current_user: TokenData = Depends(get_current_user)):
    """
    Obtener informacion del usuario autenticado
    TODO: Integrar con Azure SQL Hyperscale
    """
    # TODO: Obtener informacion completa del usuario desde la base de datos
    return User(
        id=current_user.user_id,
        email="user@example.com",
        username=current_user.username,
        full_name="User Name",
        is_active=True,
        created_at=datetime.utcnow(),
        updated_at=datetime.utcnow(),
        roles=current_user.roles
    )

# Endpoints de Usuarios
@app.get(f"/api/{API_VERSION}/users", response_model=List[User])
async def list_users(
    skip: int = 0,
    limit: int = 100,
    current_user: TokenData = Depends(get_current_user)
):
    """
    Listar usuarios (requiere autenticacion)
    TODO: Integrar con Azure SQL Hyperscale
    """
    # TODO: Implementar paginacion desde Azure SQL Hyperscale
    return []

@app.get("/api/" + API_VERSION + "/users/{user_id}", response_model=User)
async def get_user(user_id: int, current_user: TokenData = Depends(get_current_user)):
    """
    Obtener usuario por ID
    TODO: Integrar con Azure SQL Hyperscale
    """
    # TODO: Buscar usuario en Azure SQL Hyperscale
    raise HTTPException(status_code=404, detail="User not found")

# Endpoints de Roles
@app.get(f"/api/{API_VERSION}/roles", response_model=List[Role])
async def list_roles(current_user: TokenData = Depends(get_current_user)):
    """
    Listar roles disponibles
    TODO: Integrar con Azure SQL Hyperscale
    """
    return []

@app.post("/api/" + API_VERSION + "/users/{user_id}/roles/{role_id}")
async def assign_role(
    user_id: int,
    role_id: int,
    current_user: TokenData = Depends(get_current_user)
):
    """
    Asignar rol a usuario
    TODO: Integrar con Azure SQL Hyperscale
    """
    # TODO: Verificar permisos del usuario actual
    # TODO: Asignar rol en Azure SQL Hyperscale
    return {"message": "Role assigned successfully"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8001)
