backend/identity-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0
pyodbc==5.0.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-servicebus==7.11.4
prometheus-client==0.19.0
python-dotenv==1.0.0
